added check for delete by user approval

+++ src/actions/projectActions.js
@@ -38,10 +38,13 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = (id) => async dispatch => {
-    await axios.delete(`/api/project/${id}`)
-   
-    dispatch({
-        type: DELETE_PROJECT,
-        payload: id
-    });
+    if (window.confirm("Are you sure? This will delete the project and all the data related to it")) {
+        await axios.delete(`/api/project/${id}`)
+
+        dispatch({
+            type: DELETE_PROJECT,
+            payload: id
+        });
+    }
+
 };
